Added ability to select different layouts for pages.
